Refactor routes folders (especially the `auth` one)
