+++ fractal_server/app/routes/auth/__init__.py
@@ -1,0 +1,55 @@
+from fastapi_users import FastAPIUsers
+from fastapi_users.authentication import AuthenticationBackend
+from fastapi_users.authentication import BearerTransport
+from fastapi_users.authentication import CookieTransport
+from fastapi_users.authentication import JWTStrategy
+
+from fractal_server.app.models import UserOAuth
+from fractal_server.app.security import get_user_manager
+from fractal_server.config import get_settings
+from fractal_server.syringe import Inject
+
+
+bearer_transport = BearerTransport(tokenUrl="/auth/token/login")
+cookie_transport = CookieTransport(cookie_samesite="none")
+
+
+def get_jwt_strategy() -> JWTStrategy:
+    settings = Inject(get_settings)
+    return JWTStrategy(
+        secret=settings.JWT_SECRET_KEY,  # type: ignore
+        lifetime_seconds=settings.JWT_EXPIRE_SECONDS,
+    )
+
+
+def get_jwt_cookie_strategy() -> JWTStrategy:
+    settings = Inject(get_settings)
+    return JWTStrategy(
+        secret=settings.JWT_SECRET_KEY,  # type: ignore
+        lifetime_seconds=settings.COOKIE_EXPIRE_SECONDS,
+    )
+
+
+token_backend = AuthenticationBackend(
+    name="bearer-jwt",
+    transport=bearer_transport,
+    get_strategy=get_jwt_strategy,
+)
+cookie_backend = AuthenticationBackend(
+    name="cookie-jwt",
+    transport=cookie_transport,
+    get_strategy=get_jwt_cookie_strategy,
+)
+
+
+fastapi_users = FastAPIUsers[UserOAuth, int](
+    get_user_manager,
+    [token_backend, cookie_backend],
+)
+current_active_user = fastapi_users.current_user(active=True)
+current_active_verified_user = fastapi_users.current_user(
+    active=True, verified=True
+)
+current_active_superuser = fastapi_users.current_user(
+    active=True, superuser=True
+)
