+++ fractal_server/app/security/__init__.py
@@ -29,7 +29,6 @@ All routes are registerd under the `auth/` prefix.
 import contextlib
 from typing import Any
 from typing import AsyncGenerator
-from typing import Dict
 from typing import Generic
 from typing import Optional
 from typing import Type
@@ -37,12 +36,7 @@ from typing import Type
 from fastapi import Depends
 from fastapi import Request
 from fastapi_users import BaseUserManager
-from fastapi_users import FastAPIUsers
 from fastapi_users import IntegerIDMixin
-from fastapi_users.authentication import AuthenticationBackend
-from fastapi_users.authentication import BearerTransport
-from fastapi_users.authentication import CookieTransport
-from fastapi_users.authentication import JWTStrategy
 from fastapi_users.db.base import BaseUserDatabase
 from fastapi_users.exceptions import InvalidPasswordException
 from fastapi_users.exceptions import UserAlreadyExists
@@ -55,13 +49,11 @@ from sqlalchemy.orm import selectinload
 from sqlmodel import func
 from sqlmodel import select
 
-from ...config import get_settings
-from ...syringe import Inject
 from ..db import get_async_db
-from fractal_server.app.models.linkusergroup import LinkUserGroup
-from fractal_server.app.models.security import OAuthAccount
-from fractal_server.app.models.security import UserGroup
-from fractal_server.app.models.security import UserOAuth as User
+from fractal_server.app.models import LinkUserGroup
+from fractal_server.app.models import OAuthAccount
+from fractal_server.app.models import UserGroup
+from fractal_server.app.models import UserOAuth
 from fractal_server.app.schemas.user import UserCreate
 from fractal_server.logger import get_logger
 
@@ -130,7 +122,7 @@ class SQLModelUserDatabaseAsync(Generic[UP, ID], BaseUserDatabase[UP, ID]):
             return user
         return None
 
-    async def create(self, create_dict: Dict[str, Any]) -> UP:
+    async def create(self, create_dict: dict[str, Any]) -> UP:
         """Create a user."""
         user = self.user_model(**create_dict)
         self.session.add(user)
@@ -138,7 +130,7 @@ class SQLModelUserDatabaseAsync(Generic[UP, ID], BaseUserDatabase[UP, ID]):
         await self.session.refresh(user)
         return user
 
-    async def update(self, user: UP, update_dict: Dict[str, Any]) -> UP:
+    async def update(self, user: UP, update_dict: dict[str, Any]) -> UP:
         for key, value in update_dict.items():
             setattr(user, key, value)
         self.session.add(user)
@@ -151,7 +143,7 @@ class SQLModelUserDatabaseAsync(Generic[UP, ID], BaseUserDatabase[UP, ID]):
         await self.session.commit()
 
     async def add_oauth_account(
-        self, user: UP, create_dict: Dict[str, Any]
+        self, user: UP, create_dict: dict[str, Any]
     ) -> UP:  # noqa
         if self.oauth_account_model is None:
             raise NotImplementedError()
@@ -165,7 +157,7 @@ class SQLModelUserDatabaseAsync(Generic[UP, ID], BaseUserDatabase[UP, ID]):
         return user
 
     async def update_oauth_account(
-        self, user: UP, oauth_account: OAP, update_dict: Dict[str, Any]
+        self, user: UP, oauth_account: OAP, update_dict: dict[str, Any]
     ) -> UP:
         if self.oauth_account_model is None:
             raise NotImplementedError()
@@ -181,11 +173,11 @@ class SQLModelUserDatabaseAsync(Generic[UP, ID], BaseUserDatabase[UP, ID]):
 async def get_user_db(
     session: AsyncSession = Depends(get_async_db),
 ) -> AsyncGenerator[SQLModelUserDatabaseAsync, None]:
-    yield SQLModelUserDatabaseAsync(session, User, OAuthAccount)
+    yield SQLModelUserDatabaseAsync(session, UserOAuth, OAuthAccount)
 
 
-class UserManager(IntegerIDMixin, BaseUserManager[User, int]):
-    async def validate_password(self, password: str, user: User) -> None:
+class UserManager(IntegerIDMixin, BaseUserManager[UserOAuth, int]):
+    async def validate_password(self, password: str, user: UserOAuth) -> None:
         # check password length
         min_length = 4
         max_length = 100
@@ -199,7 +191,7 @@ class UserManager(IntegerIDMixin, BaseUserManager[User, int]):
             )
 
     async def on_after_register(
-        self, user: User, request: Optional[Request] = None
+        self, user: UserOAuth, request: Optional[Request] = None
     ):
         logger.info(
             f"New-user registration completed ({user.id=}, {user.email=})."
@@ -232,53 +224,6 @@ async def get_user_manager(
 ) -> AsyncGenerator[UserManager, None]:
     yield UserManager(user_db)
 
-
-bearer_transport = BearerTransport(tokenUrl="/auth/token/login")
-cookie_transport = CookieTransport(cookie_samesite="none")
-
-
-def get_jwt_strategy() -> JWTStrategy:
-    settings = Inject(get_settings)
-    return JWTStrategy(
-        secret=settings.JWT_SECRET_KEY,  # type: ignore
-        lifetime_seconds=settings.JWT_EXPIRE_SECONDS,
-    )
-
-
-def get_jwt_cookie_strategy() -> JWTStrategy:
-    settings = Inject(get_settings)
-    return JWTStrategy(
-        secret=settings.JWT_SECRET_KEY,  # type: ignore
-        lifetime_seconds=settings.COOKIE_EXPIRE_SECONDS,
-    )
-
-
-token_backend = AuthenticationBackend(
-    name="bearer-jwt",
-    transport=bearer_transport,
-    get_strategy=get_jwt_strategy,
-)
-cookie_backend = AuthenticationBackend(
-    name="cookie-jwt",
-    transport=cookie_transport,
-    get_strategy=get_jwt_cookie_strategy,
-)
-
-
-fastapi_users = FastAPIUsers[User, int](
-    get_user_manager,
-    [token_backend, cookie_backend],
-)
-
-
-# Create dependencies for users
-current_active_user = fastapi_users.current_user(active=True)
-current_active_verified_user = fastapi_users.current_user(
-    active=True, verified=True
-)
-current_active_superuser = fastapi_users.current_user(
-    active=True, superuser=True
-)
 
 get_async_session_context = contextlib.asynccontextmanager(get_async_db)
 get_user_db_context = contextlib.asynccontextmanager(get_user_db)
@@ -320,7 +265,9 @@ async def _create_first_user(
 
             if is_superuser is True:
                 # If a superuser already exists, exit
-                stm = select(User).where(User.is_superuser == True)  # noqa
+                stm = select(UserOAuth).where(
+                    UserOAuth.is_superuser == True  # noqa
+                )
                 res = await session.execute(stm)
                 existing_superuser = res.scalars().first()
                 if existing_superuser is not None:
